Admin kayıt formundaki kayıt hatası giderildi. Ve admin kaydı sonrası açılacak kullanıcı giriş formu tasarlanmaya başlandı.

+++ TRWORKFLOW/TRWORKFLOW.Entities/Concrete/User.cs
@@ -9,7 +9,7 @@ namespace TRWORKFLOW.Entities.Concrete
 {
     public class User
     {
-        public int ID { get; set; }
+        public decimal ID { get; set; }
         public string? FirstName { get; set; }
         public string? LastName { get; set; }
         public string? Mail { get; set; }
